Document polling for HelmOps (#262)
* Document semantic versioning for HelmOps

That the HelmOp version field now supports semver constraints, which
this documents with links to the spec and to examples of how such
constraints are written and parsed.

* Document polling for HelmOps

The new polling feature for HelmOps is now documented, for users to know
how to enable it and what to expect when it is enabled.

* Clarify HelmOps vs HelmOp

Feedback suggested that the distinction between HelmOps as an idea, and
`HelmOp` as a resource, needed to be expanded upon. This is an attempt
in that direction.